[Paylater v3] Removed old code and adapter viewwholder structure setup

+++ features/fintech/paylater/src/main/java/com/tokopedia/pdpsimulation/paylater/domain/usecase/PayLaterSimulationV3UseCase.kt
@@ -15,7 +15,7 @@ class PayLaterSimulationV3UseCase @Inject constructor(graphqlRepository: Graphql
     fun getPayLaterSimulationDetails(
         onSuccess: (PayLaterGetSimulation) -> Unit,
         onError: (Throwable) -> Unit,
-        amount: Long,
+        amount: Double,
         productId: String
     ) {
         try {
@@ -35,10 +35,10 @@ class PayLaterSimulationV3UseCase @Inject constructor(graphqlRepository: Graphql
     }
 
 
-    private fun getRequestParams(amount: Long, productId: String): MutableMap<String, Any?> {
+    private fun getRequestParams(amount: Double, productId: String): MutableMap<String, Any?> {
         return mutableMapOf(
             "request" to mutableMapOf(
-                PARAM_PRODUCT_AMOUNT to amount.toDouble(),
+                PARAM_PRODUCT_AMOUNT to amount,
                 PARAM_PRODUCT_ID to "1"
             )
         )
